Never actually added my design files...
This was intended to be an exciting FPGA + RISC-V board in the Raspberry Pi form factor. I nearly completed the design with a Xilinx FPGA, but decided to switch to a Lattice iCE40 due to the new open source Project Icestorm toolchain. I don't think I finished the board design, but figure I ought to add my design files :)